Move components into a data file
Because the components used a common template (which has since
been broken out into an include), it made sense to build out a
data file that could hold the information for each component.

You can find more information about Jekyll data files at:
http://jekyllrb.com/docs/datafiles/

This also redesigns the look of components so the title is above
the image, and useful links (if available) are displayed beneath
the title. Each component can now link to the documentation for it,
as well as a page with examples for how to use the component. Right
now the components which have examples have them on the same page
as the documentation, but this will likely change in the future.